sidd/requirements.txt
openai-whisper==20231117
websockets==12.0
numpy==1.24.3 